big manual merge from overwrite branch
Former-commit-id: f1abfa240bf3548e260a13c9b93922dd65799e8f

+++ trunk/src/common/metadata.hh
@@ -29,7 +29,7 @@ struct SegmentTransferOsdInfo {
 struct FileMetaData {
 	string _path;
 	uint32_t _id;
-	uint32_t _size;
+	uint64_t _size;
 	vector<uint64_t> _segmentList;
 	vector<uint32_t> _primaryList;
